+++ flink-python/pyflink/ml/tests/__init__.py
@@ -15,44 +15,3 @@
 #  See the License for the specific language governing permissions and
 # limitations under the License.
 ################################################################################
-
-import glob
-import os
-import unittest
-
-from pyflink.testing.test_case_utils import PyFlinkTestCase
-from pyflink.find_flink_home import _find_flink_source_root
-
-
-class MLTestCase(PyFlinkTestCase):
-    """
-    Base class for testing ML.
-    """
-
-    _inited = False
-
-    @staticmethod
-    def _ensure_path(pattern):
-        if not glob.glob(pattern):
-            raise unittest.SkipTest(
-                "'%s' is not available. Will skip the related tests." % pattern)
-
-    @classmethod
-    def _ensure_initialized(cls):
-        if MLTestCase._inited:
-            return
-
-        FLINK_SOURCE_ROOT_DIR = _find_flink_source_root()
-        api_path_pattern = (
-            "flink-ml-parent/flink-ml-api/target/flink-ml-api*-SNAPSHOT.jar")
-        lib_path_pattern = (
-            "flink-ml-parent/flink-ml-lib/target/flink-ml-lib*-SNAPSHOT.jar")
-
-        MLTestCase._ensure_path(os.path.join(FLINK_SOURCE_ROOT_DIR, api_path_pattern))
-        MLTestCase._ensure_path(os.path.join(FLINK_SOURCE_ROOT_DIR, lib_path_pattern))
-
-        MLTestCase._inited = True
-
-    def setUp(self):
-        super(MLTestCase, self).setUp()
-        MLTestCase._ensure_initialized()
