+++ flink-python/pyflink/testing/test_case_utils.py
@@ -15,6 +15,7 @@
 #  See the License for the specific language governing permissions and
 # limitations under the License.
 #################################################################################
+import glob
 import logging
 import os
 import re
@@ -30,7 +31,7 @@ from py4j.protocol import Py4JJavaError
 from pyflink.table.sources import CsvTableSource
 from pyflink.dataset import ExecutionEnvironment
 from pyflink.datastream import StreamExecutionEnvironment
-from pyflink.find_flink_home import _find_flink_home
+from pyflink.find_flink_home import _find_flink_home, _find_flink_source_root
 from pyflink.table import BatchTableEnvironment, StreamTableEnvironment, EnvironmentSettings
 from pyflink.java_gateway import get_gateway
 
@@ -287,3 +288,37 @@ class TestEnv(object):
         for item in self.result:
             result[item.f0] = item.f1
         return result
+
+
+class MLTestCase(PyFlinkTestCase):
+    """
+    Base class for testing ML.
+    """
+
+    _inited = False
+
+    @staticmethod
+    def _ensure_path(pattern):
+        if not glob.glob(pattern):
+            raise unittest.SkipTest(
+                "'%s' is not available. Will skip the related tests." % pattern)
+
+    @classmethod
+    def _ensure_initialized(cls):
+        if MLTestCase._inited:
+            return
+
+        flink_source_root_dir = _find_flink_source_root()
+        api_path_pattern = (
+            "flink-ml-parent/flink-ml-api/target/flink-ml-api*-SNAPSHOT.jar")
+        lib_path_pattern = (
+            "flink-ml-parent/flink-ml-lib/target/flink-ml-lib*-SNAPSHOT.jar")
+
+        MLTestCase._ensure_path(os.path.join(flink_source_root_dir, api_path_pattern))
+        MLTestCase._ensure_path(os.path.join(flink_source_root_dir, lib_path_pattern))
+
+        MLTestCase._inited = True
+
+    def setUp(self):
+        super(MLTestCase, self).setUp()
+        MLTestCase._ensure_initialized()
